+++ http-client/Network/HTTP/Client/Connection.hs
@@ -41,7 +41,7 @@ connectionReadLineWith conn bs0 =
     go bs0 id 0
   where
     go bs front total =
-        case S.breakByte charLF bs of
+        case S.break (== charLF) bs of
             (_, "") -> do
                 let total' = total + S.length bs
                 when (total' > 4096) $ throwIO OverlongHeaders
@@ -85,15 +85,39 @@ makeConnection :: IO ByteString -- ^ read
                -> IO Connection
 makeConnection r w c = do
     istack <- newIORef []
+
+    -- it is necessary to make sure we never read from or write to
+    -- already closed connection.
+    closedVar <- newIORef False
+
     _ <- mkWeakIORef istack c
     return $! Connection
-        { connectionRead = join $ atomicModifyIORef istack $ \stack ->
-            case stack of
-                x:xs -> (xs, return x)
-                [] -> ([], r)
-        , connectionUnread = \x -> atomicModifyIORef istack $ \stack -> (x:stack, ())
-        , connectionWrite = w
-        , connectionClose = c
+        { connectionRead = do
+            closed <- readIORef closedVar
+            when closed $
+              throwIO ConnectionClosed
+            join $ atomicModifyIORef istack $ \stack ->
+              case stack of
+                  x:xs -> (xs, return x)
+                  [] -> ([], r)
+
+        , connectionUnread = \x -> do
+            closed <- readIORef closedVar
+            when closed $
+              throwIO ConnectionClosed
+            atomicModifyIORef istack $ \stack -> (x:stack, ())
+
+        , connectionWrite = \x -> do
+            closed <- readIORef closedVar
+            when closed $
+              throwIO ConnectionClosed
+            w x
+
+        , connectionClose = do
+            closed <- readIORef closedVar
+            unless closed $
+              c
+            writeIORef closedVar True
         }
 
 socketConnection :: Socket -> Int -> IO Connection
@@ -141,8 +165,8 @@ openSocketConnectionSize tweakSocket chunksize hostAddress host port = do
             (NS.sClose)
             (\sock -> do
                 NS.setSocketOption sock NS.NoDelay 1
-                NS.connect sock (NS.addrAddress addr)
                 tweakSocket sock
+                NS.connect sock (NS.addrAddress addr)
                 socketConnection sock chunksize)
 
 firstSuccessful :: [NS.AddrInfo] -> (NS.AddrInfo -> IO a) -> IO a