+++ http-client/Network/HTTP/Client/Manager.hs
@@ -22,6 +22,9 @@ module Network.HTTP.Client.Manager
     , dropProxyAuthSecure
     ) where
 
+#ifndef MIN_VERSION_base
+#define MIN_VERSION_base(x,y,z) 1
+#endif
 #if !MIN_VERSION_base(4,6,0)
 import Prelude hiding (catch)
 #endif
@@ -37,6 +40,7 @@ import qualified Data.ByteString.Lazy as L
 
 import qualified Blaze.ByteString.Builder as Blaze
 
+import Data.Char (toLower)
 import Data.Text (Text)
 import qualified Data.Text as T
 import Data.Text.Read (decimal)
@@ -57,7 +61,7 @@ import Network.HTTP.Types (status200)
 import Network.HTTP.Client.Types
 import Network.HTTP.Client.Connection
 import Network.HTTP.Client.Headers (parseStatusHeaders)
-import Network.HTTP.Client.Request (username, password, applyBasicProxyAuth)
+import Network.HTTP.Client.Request (applyBasicProxyAuth, extractBasicAuthInfo)
 import Control.Concurrent.MVar (MVar, takeMVar, tryPutMVar, newEmptyMVar)
 import System.Environment (getEnvironment)
 import qualified Network.URI as U
@@ -149,7 +153,7 @@ putSocket man key ci = do
                  in m' `seq` (m', return ())
             Just l ->
                 let (l', mx) = addToList now (mMaxConns man) ci l
-                    cnt' = idleCount + maybe 0 (const 1) mx
+                    cnt' = idleCount + maybe 1 (const 0) mx
                     m' = ManagerOpen cnt' (Map.insert key l' m)
                  in m' `seq` (m', maybe (return ()) connectionClose mx)
 
@@ -430,6 +434,7 @@ getConn req m
                 let ultHost = host req
                     ultPort = port req
                     proxyAuthorizationHeader = maybe "" (\h -> S8.concat ["Proxy-Authorization: ", h, "\r\n"]) . lookup "Proxy-Authorization" $ requestHeaders req
+                    hostHeader = S8.concat ["Host: ", ultHost, (S8.pack $ show ultPort), "\r\n"]
                     connstr = S8.concat
                         [ "CONNECT "
                         , ultHost
@@ -437,12 +442,13 @@ getConn req m
                         , S8.pack $ show ultPort
                         , " HTTP/1.1\r\n"
                         , proxyAuthorizationHeader
+                        , hostHeader
                         , "\r\n"
                         ]
                     parse conn = do
                         sh@(StatusHeaders status _ _) <- parseStatusHeaders conn Nothing Nothing
                         unless (status == status200) $
-                            throwIO $ ProxyConnectException ultHost ultPort $ Left $ S8.pack $ show sh
+                            throwIO $ ProxyConnectException ultHost ultPort $ Right $ StatusCodeException status [] (CJ [])
                  in mTlsProxyConnection m connstr parse (S8.unpack ultHost)
 
 -- | Get the proxy settings from the @Request@ itself.
@@ -506,7 +512,9 @@ envHelper :: Text -> EnvHelper -> IO (Request -> Request)
 envHelper name eh = do
     env <- getEnvironment
     let lenv = Map.fromList $ map (first $ T.toLower . T.pack) env
-    case lookup (T.unpack name) env <|> Map.lookup name lenv of
+        lookupEnvVar n = lookup (T.unpack n) env <|> Map.lookup n lenv
+        noProxyDomains = domainSuffixes (lookupEnvVar "no_proxy")
+    case lookupEnvVar name of
         Nothing  -> return noEnvProxy
         Just ""  -> return noEnvProxy
         Just str -> do
@@ -522,14 +530,6 @@ envHelper name eh = do
                 guard $ null $ U.uriFragment uri
 
                 auth <- U.uriAuthority uri
-                let muserpass =
-                        if null authInfo
-                            then Nothing
-                            else Just ( S8.pack $ username authInfo
-                                      , S8.pack $ password authInfo
-                                      )
-                    authInfo = U.uriUserInfo auth
-
                 port <-
                     case U.uriPort auth of
                         "" -> Just 80
@@ -539,11 +539,19 @@ envHelper name eh = do
                                 _ -> Nothing
                         _ -> Nothing
 
-                Just $ (Proxy (S8.pack $ U.uriRegName auth) port, muserpass)
+                Just $ (Proxy (S8.pack $ U.uriRegName auth) port, extractBasicAuthInfo uri)
             return $ \req ->
-                maybe id (uncurry applyBasicProxyAuth) muserpass
-                req { proxy = Just p }
+                if host req `hasDomainSuffixIn` noProxyDomains
+                then noEnvProxy req
+                else maybe id (uncurry applyBasicProxyAuth) muserpass
+                     req { proxy = Just p }
     where noEnvProxy = case eh of
             EHFromRequest -> id
             EHNoProxy     -> \req -> req { proxy = Nothing }
             EHUseProxy p  -> \req -> req { proxy = Just p  }
+          prefixed s | S8.head s == '.' = s
+                     | otherwise = S8.cons '.' s
+          domainSuffixes Nothing = []
+          domainSuffixes (Just "") = []
+          domainSuffixes (Just no_proxy) = [prefixed $ S8.dropWhile (== ' ') suffix | suffix <- S8.split ',' (S8.pack (map toLower no_proxy)), not (S8.null suffix)]
+          hasDomainSuffixIn host = any (`S8.isSuffixOf` prefixed (S8.map toLower host))