+++ http-conduit/Network/HTTP/Client/Conduit.hs
@@ -3,8 +3,13 @@
 {-# LANGUAGE RankNTypes            #-}
 -- | A new, experimental API to replace "Network.HTTP.Conduit".
 --
--- For more information, please be sure to read the documentation in the
--- "Network.HTTP.Client" module.
+-- For most users, "Network.HTTP.Simple" is probably a better choice. For more
+-- information, see:
+--
+-- <https://github.com/commercialhaskell/jump/blob/master/doc/http-client.md>.
+--
+-- For more information on using this module, please be sure to read the
+-- documentation in the "Network.HTTP.Client" module.
 module Network.HTTP.Client.Conduit
     ( -- * Conduit-specific interface
       withResponse
@@ -17,7 +22,6 @@ module Network.HTTP.Client.Conduit
     , withManager
     , withManagerSettings
     , newManagerSettings
-    , HasHttpManager (..)
       -- * General HTTP client interface
     , module Network.HTTP.Client
     , httpLbs
@@ -45,6 +49,7 @@ import           Network.HTTP.Client          hiding (closeManager,
                                                newManager, responseClose,
                                                responseOpen, withManager,
                                                withResponse, BodyReader, brRead, brConsume, httpNoBody)
+import           Network.HTTP.Client          (HasHttpManager(..))
 import qualified Network.HTTP.Client          as H
 import           Network.HTTP.Client.TLS      (tlsManagerSettings)
 
@@ -125,11 +130,6 @@ responseOpen req = do
 -- Since 2.1.0
 responseClose :: MonadIO m => Response body -> m ()
 responseClose = liftIO . H.responseClose
-
-class HasHttpManager a where
-    getHttpManager :: a -> Manager
-instance HasHttpManager Manager where
-    getHttpManager = id
 
 bodyReaderSource :: MonadIO m
                  => H.BodyReader