+++ http-conduit/ChangeLog.md
@@ -1,3 +1,23 @@
+## 2.1.11
+
+* Switch to non-throwing behavior in `Network.HTTP.Simple` [#193](https://github.com/snoyberg/http-client/issues/193)
+
+## 2.1.10.1
+
+* Fix mistaken `@since` comments
+
+## 2.1.10
+
+* Add the `Network.HTTP.Simple` module
+
+## 2.1.9
+
+* cabal file cleanup
+
+## 2.1.8
+
+* Move HasHttpManager from http-conduit to http-client [#147](https://github.com/snoyberg/http-client/pull/147)
+
 ## 2.1.7
 
 * Deprecate `conduitManagerSettings`, re-export `tlsManagerSettings` [#136](https://github.com/snoyberg/http-client/issues/136) [#137](https://github.com/snoyberg/http-client/issues/137)